feat(backend): 2値ポイント + 役職傾斜 + 集計バッジ処理
ポイント体系を 9セル × 3アクション × 3カテゴリ から
2値（daily 0.1P / creative 5P）× 3カテゴリ + 役職傾斜 に再設計。

主な変更:
- PointApplication に level / category / final_point カラム追加
- points 列を Integer → Numeric(10, 2) （daily 0.1P を欠落させない）
- 役職傾斜: 課長以上 = 3.0x / それ以外 = 1.0x （app/services/point_calc.py）
- PointsInput を 9キー → 3キー (social/safety/future) に集約
- calculator の M9_7 から M3_7 を平均集約（option A、6,000P で目標達成は維持）
- aggregate_approved_points を level/category ベースに書き換え
- cascade router から旧 5→9 暫定変換を撤廃
- point_applications router で level + category 受理、
  activity_genre_id を自動導出して dashboard / mypage 互換維持
- department_category_points テーブル + バッチ集計スクリプト追加
- Alembic migration c83e1f4a2d09:
  ・新カラム追加 + 越境×〜 row 削除
  ・既存 row の level/category/final_point を役職傾斜つきで再計算
  ・activity_genres から越境×〜 を削除
  ・department_category_points テーブル新設
- 6 ジャンル版 seed_activity_genres + 全 seed スクリプト更新

5/14 きみなり方針共有・5/16 確定事項を反映。
refs #2 #17

+++ backend/app/services/dashboard_service.py
@@ -233,7 +233,7 @@ def aggregate_points_summary(
         or 0
     )
 
-    return {"total": int(total), "annual_target": annual_target}
+    return {"total": float(total), "annual_target": annual_target}
 
 
 # ════════════════════════════════════════════════════════════
@@ -281,10 +281,10 @@ def aggregate_my_points_by_genre(
             "activity_genre_id": int(gid),
             "activity_genre_name": str(gname),
             "sort_order": int(sort_order),
-            "points": int(pts),
+            "points": float(pts),
         }
         for gid, gname, sort_order, pts in rows_raw
-        if int(pts) != 0
+        if float(pts) != 0.0
     ]
     total = sum(r["points"] for r in rows)
     return {"rows": rows, "total_points": total}
@@ -329,8 +329,8 @@ def aggregate_org_member_points_by_genre(
     )
     out: list[dict] = []
     for uid, uname, gid, gname, sort_order, pts in session.execute(stmt).all():
-        p = int(pts)
-        if p == 0:
+        p = float(pts)
+        if p == 0.0:
             continue
         out.append(
             {
